Project/PWA trial/PROJECT/BACKEND/requirements.txt
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.2.0
Flask-SQLAlchemy==3.0.3
Flask-SSE==1.0.0
Flask-WTF==1.1.1
numpy==1.23.2
pandas==1.5.3
matplotlib==3.7.0
Jinja2==3.1.2
jsonschema==4.17.3
gunicorn==21.2.0
httplib2==0.22.0
gevent==23.7.0
gitdb==4.0.10
GitPython==3.1.31
email_validator
bcrypt==4.0.1
flask_restful
redis
celery[redis]
async-generator==1.10
async-timeout==4.0.2
PyHTML==1.3.2
PyJWT==2.8.0
requests==2.28.2
WeasyPrint==52.5
urllib3==1.26.14